+++ CHANGELOG.rst
@@ -35,6 +35,8 @@ Added
 
 - New Failover API function ``set_options`` to change failover internal params.
 
+- Issue about sync spaces usage with a wrong failover setup.
+
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Changed
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~